Renamed Schema and SchemaRoot actors to PartitionCreator and Schema, adapted their names and dispatchers accordingly, changed guardian supervision strategy
