+++ schedoscope-core/src/main/resources/reference.conf
@@ -628,6 +628,8 @@ akka {
     
     actor {        
     
+    	guardian-supervisor-strategy = "akka.actor.StoppingSupervisorStrategy"
+    
         #
         # Given the variety of component-specific dispatchers used by Schedoscope
         # we limit the Akka default dispatcher to 8 threads.
@@ -684,7 +686,7 @@ akka {
         # changed.
         #
 
-        schema-root-actor-dispatcher {
+        schema-actor-dispatcher {
             executor = "thread-pool-executor"
             type = PinnedDispatcher         
         }           
@@ -695,7 +697,7 @@ akka {
         # be aligned with the property schedoscope.metastore.concurrency 
         #
         
-        schema-actor-dispatcher {
+        partition-creator-dispatcher {
             executor = "thread-pool-executor"
             type = Dispatcher
