+++ schedoscope-core/src/main/scala/org/schedoscope/Schedoscope.scala
@@ -46,19 +46,19 @@ object Schedoscope {
   lazy val viewManagerActor = actorSelectionToRef(actorSystem.actorSelection(rootActor.path.child("views"))).get
 
   /**
-   * A reference to the Schedoscope schema root actor
-   */
-  lazy val schemaRootActor = actorSelectionToRef(actorSystem.actorSelection(rootActor.path.child("schema-root"))).get
-
-  /**
    * A reference to the Schedoscope schema actor
    */
-  lazy val schemaActor = actorSelectionToRef(actorSystem.actorSelection(schemaRootActor.path.child("schema"))).get
+  lazy val schemaActor = actorSelectionToRef(actorSystem.actorSelection(rootActor.path.child("schema"))).get
+
+  /**
+   * A reference to the Schedoscope partition creator actor
+   */
+  lazy val partitionCreatorActor = actorSelectionToRef(actorSystem.actorSelection(schemaActor.path.child("partition-creator"))).get
 
   /**
    * A reference to the Schedoscope metadata logger actor
    */
-  lazy val metadataLoggerActor = actorSelectionToRef(actorSystem.actorSelection(schemaRootActor.path.child("metadata-logger"))).get
+  lazy val metadataLoggerActor = actorSelectionToRef(actorSystem.actorSelection(schemaActor.path.child("metadata-logger"))).get
 
   /**
    * A reference to the Schedoscope transformation manager logger actor
